Add files via upload
The form-group class collects user input for grid shape, rows and columns. The gridContainer holds the generated grid and the sidebar displays the selected elements in a collapsable list. 
The Form Submission Handling calls the generateGrid function to create the grid based on user input. The grid generator retrieves user input for shape, rows, and columns and clears the existing grid. 
Challenges: getting the triangular and ellipse shapes.